+++ libs/common/src/abstractions/import/import.service.abstraction.ts
@@ -1,6 +1,6 @@
-import { ImportOption, ImportType } from "../enums/importOptions";
-import { ImportError } from "../importers/import-error";
-import { Importer } from "../importers/importer";
+import { ImportOption, ImportType } from "../../enums/importOptions";
+import { ImportError } from "../../importers/import-error";
+import { Importer } from "../../importers/importer";
 
 export abstract class ImportService {
   featuredImportOptions: readonly ImportOption[];
